Move MinMap to utils
