+++ application/src/main/java/org/opentripplanner/graph_builder/module/nearbystops/NearbyStopFinderVisitor.java
@@ -11,7 +11,6 @@ import java.util.Set;
 import org.opentripplanner.astar.spi.TraverseVisitor;
 import org.opentripplanner.core.model.id.FeedScopedId;
 import org.opentripplanner.framework.application.OTPFeature;
-import org.opentripplanner.graph_builder.module.transfer.filter.MinMap;
 import org.opentripplanner.routing.graphfinder.NearbyStop;
 import org.opentripplanner.street.model.edge.Edge;
 import org.opentripplanner.street.model.edge.StreetEdge;
@@ -20,6 +19,7 @@ import org.opentripplanner.street.model.vertex.TransitStopVertex;
 import org.opentripplanner.street.model.vertex.Vertex;
 import org.opentripplanner.street.search.TraverseMode;
 import org.opentripplanner.street.search.state.State;
+import org.opentripplanner.utils.collection.MinMap;
 
 /**
  * A {@link TraverseVisitor} that collects transit stops and flex area stops during an A* search,
